requirements.txt
torch
torchvision
mmcv-full
wandb
prettytable
